+++ NEWS.md
@@ -23,6 +23,8 @@
 - New file `spatial_manipulation.R`
 - Add `ext()` methods for `giottoPolygon`, `giottoPoints`, `spatialNetworkObj`, `spatLocsObj`
 - Add `flip()` methods for `giottoPolygon`, `giottoPoints`, `spatialNetworkObj`, `spatLocsObj`, `SpatExtent`
+- Add access to terra plotting params through ... param for `giottoLargeImage` `plot()` method.
+- Add params for `maxcell` and `col` for `giottoLargeImage` `plot()` method that respectively control how many image cells max are used and what colors are used when plotting single-channel images.
 
 ## Changes
 - Fix bug in `combine_matrices()`
@@ -46,6 +48,7 @@
 - Update cell segmentation workflow to check for *deepcell* and *PIL* python packages
 - Update cell segmentation workflow to return grayscale mask images instead of RGB
 - Update `createGiottoVisiumObject()` image h5 scalefactors reading to use partial matching for whether hi or lowres image is supplied
+- Update `giottoLargeImage` `plot()` method to use `terra::plot()` instead of `terra::plotRGB()` for grayscale images
 
 
 # Giotto Suite 3.3.0 (2023-04-18)